refactor: Move config modules to folder and hoist logic
