+++ cypress/tsconfig.json
@@ -7,7 +7,7 @@
   "include": [
     "**/*", // override the include from the root tsconfig
     "../src/internalTypes.d.ts", // contains definitions for global variables
-    "../src/playroomModuleAliases.d.ts"
+    "../src/configModules/configModules.d.ts"
   ],
   "exclude": [] // override the exclude from the root tsconfig
 }
