--- src/components.ts
@@ -1,4 +1,0 @@
-import * as components from '__PLAYROOM_ALIAS__COMPONENTS__';
-
-// eslint-disable-next-line @typescript-eslint/consistent-type-imports
-export default components as typeof import('__PLAYROOM_ALIAS__COMPONENTS__').default;
