+++ src/configModules/configModules.d.ts
