+++ src/configModules/frameComponent.ts
