+++ src/configModules/useScope.ts
@@ -1,0 +1,33 @@
+import userScopeFromConfig from '__PLAYROOM_ALIAS__USE_SCOPE__';
+import React, { createElement, Fragment } from 'react';
+
+import {
+  ReactCreateElementPragma,
+  ReactFragmentPragma,
+} from '../utils/compileJsx';
+
+import components from './components';
+
+const userScope = {
+  ...(userScopeFromConfig() ?? {}),
+  ...components,
+};
+
+if (ReactCreateElementPragma in userScope) {
+  throw new Error(
+    `'${ReactCreateElementPragma}' is used internally by Playroom and is not allowed in scope`
+  );
+}
+
+if (ReactFragmentPragma in userScope) {
+  throw new Error(
+    `'${ReactFragmentPragma}' is used internally by Playroom and is not allowed in scope`
+  );
+}
+
+export default {
+  ...userScope,
+  React,
+  [ReactCreateElementPragma]: createElement,
+  [ReactFragmentPragma]: Fragment,
+};
