+++ src/entries/preview.tsx
@@ -3,11 +3,7 @@ import faviconPath from '../../images/favicon.png';
 import Frame from '../Playroom/Frame/Frame';
 import Preview from '../Playroom/Preview/Preview';
 import { PreviewError } from '../Playroom/Preview/PreviewError';
-import playroomComponents from '../components';
-import frameComponent from '../frameComponent';
 import { renderElement } from '../render';
-import playroomThemes from '../themes';
-import { hmrAccept } from '../utils/hmr';
 import { UrlParams } from '../utils/params';
 
 const outlet = document.createElement('div');
@@ -24,41 +20,18 @@ if (selectedElement) {
   selectedElement.setAttribute('href', formattedFavicon);
 }
 
-const renderPreview = ({
-  themes = playroomThemes,
-  components = playroomComponents,
-  FrameComponent = frameComponent,
-} = {}) => {
-  renderElement(
-    <UrlParams themes={themes} decodeUrl>
-      {({ code, themeName, theme, title }) => (
-        <Preview title={title}>
-          <Frame
-            code={code}
-            components={components}
-            themeName={themeName}
-            theme={theme}
-            FrameComponent={FrameComponent}
-            ErrorComponent={PreviewError}
-          />
-        </Preview>
-      )}
-    </UrlParams>,
-    outlet
-  );
-};
-renderPreview();
-
-hmrAccept((accept) => {
-  accept('./components', () => {
-    renderPreview({ components: playroomComponents });
-  });
-
-  accept('./themes', () => {
-    renderPreview({ themes: playroomThemes });
-  });
-
-  accept('./frameComponent', () => {
-    renderPreview({ FrameComponent: frameComponent });
-  });
-});
+renderElement(
+  <UrlParams decodeUrl>
+    {({ code, themeName, theme, title }) => (
+      <Preview title={title}>
+        <Frame
+          code={code}
+          themeName={themeName}
+          theme={theme}
+          ErrorComponent={PreviewError}
+        />
+      </Preview>
+    )}
+  </UrlParams>,
+  outlet
+);
