--- src/useScope.ts
@@ -1,1 +1,0 @@
-export { default } from '__PLAYROOM_ALIAS__USE_SCOPE__';
