--- src/utils/hmr.ts
@@ -1,10 +1,0 @@
-type AcceptHandler = (relativePath: string, handler: () => void) => void;
-
-export function hmrAccept(handler: (accept: AcceptHandler) => void) {
-  if (
-    typeof module !== 'undefined' &&
-    typeof module?.hot?.accept === 'function'
-  ) {
-    handler(module.hot.accept);
-  }
-}
